feat: authorize update and destroy actions using Idea policy gates

+++ app/Http/Controllers/IdeaController.php
@@ -78,6 +78,8 @@ class IdeaController extends Controller
      */
     public function update(IdeaRequest $request, Idea $idea)
     {
+        Gate::authorize('update', $idea);
+
         $idea->update($request->validated());
 
         return redirect('/ideas');
@@ -88,6 +90,8 @@ class IdeaController extends Controller
      */
     public function destroy(Idea $idea)
     {
+        Gate::authorize('update', $idea);
+
         $idea->delete();
 
         return redirect('/ideas');
